refactor CPoolSwap composers. Extract maxPriceAsBigNumber composer to buffer_composer.ts
